requirements.txt
google-generativeai
python-dotenv
streamlit
pip
setuptools
wheel
